refactor: [#15] remove duplicate code and move file path logic to librarian domain
