+++ tests/test_nautilus_librarian/test_domain/test_file_locator.py
@@ -1,4 +1,4 @@
-from nautilus_librarian.mods.namecodes.domain.file_locator import file_locator
+from nautilus_librarian.domain.file_locator import file_locator
 from nautilus_librarian.mods.namecodes.domain.filename import Filename
