+++ tests/test_nautilus_librarian/test_mods/test_namecodes/test_domain/test_filename.py
@@ -6,6 +6,6 @@ def test_instantiation_from_string():
 
     assert gold_image_filename.artwork_id == "000001"
     assert gold_image_filename.purpose_code == "32"
-    assert gold_image_filename.transformacion_code == "600"
+    assert gold_image_filename.transformation_code == "600"
     assert gold_image_filename.type_code == "2"
     assert gold_image_filename.extension == "tif"
